+++ R/canvas_split.R
@@ -1,0 +1,64 @@
+# Copyright (C) 2021-2022 Koen Derks
+
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+#' Draw Split Lines
+#'
+#' @description This function draws split lines.
+#'
+#' @usage canvas_split(colors, background = "black", iterations = 7,
+#'              lwd = 0.05, alpha = 0.2)
+#'
+#' @param colors         a string or character vector specifying the color(s) used for the artwork.
+#' @param background     a character specifying the color used for the background (and the hole).
+#' @param iterations     a positive integer specifying the number of iterations of the algorithm.
+#' @param lwd            a numeric value specifying the width of the lines.
+#' @param alpha          a numeric value specifying the transparency of the lines.
+#'
+#' @return A \code{ggplot} object containing the artwork.
+#'
+#' @author Koen Derks, \email{koen-derks@hotmail.com}
+#'
+#' @keywords artwork canvas
+#'
+#' @seealso \code{colorPalette}
+#'
+#' @examples
+#' \donttest{
+#' set.seed(2)
+#'
+#' # Simple example
+#' canvas_split(colors = colorPalette("origami"))
+#' }
+#'
+#' @export
+
+canvas_split <- function(colors, background = "black", iterations = 7,
+                         lwd = 0.05, alpha = 0.2) {
+  .checkUserInput(iterations = iterations, background = background)
+  line <- data.frame(
+    x = c(0, 1, 1, 0), xend = c(1, 1, 0, 0),
+    y = c(0, 0, 1, 1), yend = c(0, 1, 1, 0),
+    col = sample(1:length(colors), size = 4, replace = TRUE)
+  )
+  canvas <- draw_split(line$x, line$xend, line$y, line$yend, line$col, length(colors), iterations)
+  breaks <- range(c(canvas$x, canvas$xend, canvas$y, canvas$yend))
+  p <- ggplot2::ggplot(data = canvas) +
+    ggplot2::geom_segment(mapping = ggplot2::aes(x = x, y = y, xend = xend, yend = yend, col = factor(col)), size = lwd, alpha = alpha) +
+    ggplot2::scale_x_continuous(limits = breaks) +
+    ggplot2::scale_y_continuous(limits = breaks) +
+    ggplot2::scale_color_manual(values = colors)
+  p <- aRtsy::theme_canvas(p, background = background)
+  return(p)
+}